fixed typos, changed DBFactory
