+++ neo/Settings.py
@@ -146,13 +146,13 @@ class SettingsHolder:
     def database_properties(self):
         return {'blockchain': {'path': self.chain_leveldb_path,
                                'skip_version_check': False,
-                               'backend': 'leveldb'},
+                               'backend': 'neo.Storage.Implementation.LevelDB.LevelDBImpl.LevelDBImpl'},
 
                 'notification': {'path': self.notification_leveldb_path,
-                                 'backend': 'leveldb'},
+                                 'backend': 'neo.Storage.Implementation.LevelDB.LevelDBImpl.LevelDBImpl'},
 
                 'debug': {'path': self.debug_storage_leveldb_path,
-                          'backend': 'leveldb'}
+                          'backend': 'neo.Storage.Implementation.LevelDB.LevelDBImpl.LevelDBImpl'}
                 }
 
     # Helpers
